[rest] Supports global system tables (#4880)
