+++ paimon-core/src/test/java/org/apache/paimon/table/system/AllTableOptionsTableTest.java
@@ -59,11 +59,12 @@ public class AllTableOptionsTableTest extends TableTestBase {
     }
 
     @Test
-    public void testSchemasTable() throws Exception {
+    public void testAllTableOptionsTable() throws Exception {
         List<String> result =
                 read(allTableOptionsTable).stream()
                         .map(Objects::toString)
                         .collect(Collectors.toList());
+        result = result.stream().filter(r -> !r.contains("path")).collect(Collectors.toList());
         assertThat(result)
                 .containsExactlyInAnyOrder(
                         "+I(default,T,fields.sales.aggregate-function,sum)",
